+++ CHANGES.rst
@@ -1,6 +1,25 @@
 Changelog
 =========
 
+Version 0.6
+-----------
+
+*2020-01-19*
+
+**Backward-incompatible changes:**
+
+- Removed the deprecated ``Duration.isDoubleDotted`` property.
+
+**Deprecations:**
+
+- Deprecated the usage of ``Beat.realStart`` in favor of ``Beat.startInMeasure``. The former will be removed in the next
+  release.
+
+**Changes:**
+
+- Updated `attrs <https://attrs.readthedocs.io>`_ to at least 19.2.
+
+
 Version 0.5
 -----------